Fix typo in Error handling controller code
On the frontend, the error message 'Username already exist' appears when creating a user with the same username as an existing user.  Grammatically, this error looks strange.  A more proper error message would be 'Username already exists'.  Possibly later on support can be added to properly conjugate this error message for a 'plural' field.