feat(gulp): Gulp Server Watch for Mocha
Adds a Gulp task that watches all server files (including server tests),
and upon any changes it will perform the Gulp *test:server* task.

Added a watch for server assets, to the main gulp watch config. This will only
add the watch when the NODE_ENV is set to "test".

Also, includes an **optional** argument for the task that can specify
that if the changed file is a test, then it will only run that test
file.

Example usage: gulp test:server:watch --onlyChanged