fix(dependency): Mongoose 4.4.7 bug
Fixes an issue with Mongoose `4.4.7` that causes the database
connection.db property to be undefined. This issue is causing our builds
to fail.

Sets the Mongoose version to `~4.4.3 <4.4.7` for now until we
investigate further. This should solve the issue of our failing builds,
when the `dropdb` Grunt/Gulp task is ran with the E2E tests.